+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Create: TFMG Community Edition 1.3.0
+# Create: TFMG Community Edition 1.2.4b
 ## Rant:
 Something something version idk :3
 
@@ -19,6 +19,7 @@ Contributors:<br>
   - The amount drained now matches recipe input amount.
   - The amount filled now matches the recipe result amount.
   - Added a delay between processes based on the heat level of the tank.
+- Added ponder for the Winding Machine.
 
 
 ### API Changes: